fixmatch/requirements.txt
apex==0.9.10dev
augly==1.0.0
aum==1.0.2
click==8.1.7
cloudpickle==3.0.0
loguru==0.7.2
matplotlib==3.9.0
medmnist==3.0.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
pydantic==2.7.4
pytest==8.2.2
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.5.0
seaborn==0.13.2
semilearn==0.3.2
setuptools==69.5.1
six==1.16.0
timm==1.0.3
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
